refactor: room 에 관련된 controller 분리
